Port AFSK module to mbed
I implemented an IRadioOut interface that can be used to plug-in
different methods to create audio out signal for the radio.  Currently
there is just an implementation which knows how to use mbed's AnalogOut
class.

I have also removed the buzzer and sensor code since it won't be
initially needed in the mbed version.

The mbed AFSK module accepts a buffer to send and then does the
modulation in the background on a timer.  No polling needs to be done
from the main thread.